requirements.txt
streamlit>=1.36
pandas>=2.0
numpy>=1.25
seaborn>=0.13
matplotlib>=3.7
openpyxl>=3.1
XlsxWriter>=3.1
Pillow>=10.0